requirements.txt
asyncio==3.4.3
aiogram==3.6.0
motor==3.4.0
pydantic-settings==2.2.1
python-dateutil==2.9.0